Set swappiness to reflect the zram vs disk speed
The kernel documentation explains that vm.swappiness reflects the speed
ratio between swapping and filesystem access when reloading discarded
pages). My own testing produced the same result, 180, as someone else
published on Reddit. The default of 60 is only appropriate when swapping
on the same type of device as the one hosting your filesystem.
Otherwise, as the kernel documentation indicates, the swappiness should
have values over 100. In my case, filesystem access on the Mac Mini HDD
is about 10 times slower than reading and writing to zram with zstd.

https://www.reddit.com/r/Fedora/comments/mzun99/new_zram_tuning_benchmarks/
https://www.kernel.org/doc/html/latest/admin-guide/sysctl/vm.html#swappiness